fix(ui-overlays): do not unset duration for modals
To account for an intermittent issue in IE11, the transition duration
for modal opening and closing is hard-coded to be '0ms'. However, the
way this had been handled before inadvertently set the duration as
`undefined` for all other browsers. While this did not manifest in any
tangible way for users, it meant that the transition duration in tests
would sometimes be stuck defaulting to the '300ms' value, even when
overridden when `use()`-ing the theme in test setup. This means slow
tests. Very slow.

test plan:

To ensure the intended behavior in IE11 is unaffected, the `onEntered`
and `onExited` callback props for a `Modal` will need to be set to
something observable for verification. This could be as simple as a
`console.log()` statement. If those callbacks are executed when the
`Modal` opens and closes, respectively, then everything should be fine.

Change-Id: I5654cd9d7afe1932561a172db679599c18c4ccf6
Reviewed-on: https://gerrit.instructure.com/208130
Tested-by: Jenkins
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Visual-Regression-Test: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>