fix(ui-themeable): check JSDOM environment
JSDOM doesn't support custom properties, but
we don't necessarily want to apply the custom
properties polyfill.
The polyfill adds a lot of unncessary style tags
which makes it difficult to
read the output of tests.
This change adds a check to see if the JSDOM user agent is applied. If it
is, then it will NOT apply the custom
properties polyfill, making it easier to read DOM output when
reviewing tests.
TEST PLAN:
In a library that consumes ui-themeable, execute tests with Jest that render DOM. The DOM should not
contain extra, unnecessary <style> tags.

Change-Id: I1fd7139021f2500f0ef76c09dbd31c9d4b97694a
Reviewed-on: https://gerrit.instructure.com/c/instructure-ui/+/226585
Reviewed-by: Steve Jensen <sejensen@instructure.com>
QA-Review: Steve Jensen <sejensen@instructure.com>
Product-Review: Steve Jensen <sejensen@instructure.com>
Visual-Regression-Test: Steve Jensen <sejensen@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>