fix(ui-tag): show cut off letter descenders
refs: INSTUI-2237

Expands line height for inner text element to the height of the tag,
less the width of the borders.

TEST PLAN:
- Use tag text containing one or more characters with a large ascender
or descender, and verify that the characters are fully visible, e.g.:

```
<Tag text="ʃ̬ long text to test for truncation" />
```

- Inspect the vertical alignment of text within the tag for all themes.

NOTE: Due to differences in font design, the vertical alignment will not
be the same for all font families. For example, the "instructure" theme
uses Proxima Nova, so the baseline will appear to be shifted up slightly
when compared to the Lato font family used for "canvas" themes. This
effect is also visible in other components, such as
[Pill](https://instructure.design/#Pill)

Change-Id: I6199a0f4f963341ffdd917ed8b612bcfff37bb8e
Reviewed-on: https://gerrit.instructure.com/c/instructure-ui/+/228551
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Pam Hiett <phiett@instructure.com>
Product-Review: Pam Hiett <phiett@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Visual-Regression-Test: Steve Jensen <sejensen@instructure.com>