chore(*): restore cjs output
Jest does not have great support for es modules atm. There are possible workarounds,
but nothing that is straightforward. We don't want to pass on that complexity to
the consumer when the infrastructure is already in place on the instui side and can
be solved trivially on our end. In the future when there is more de facto es module
support, we can drop the cjs output easily by reverting this change.

refs: INSTUI-2693

TEST PLAN:
- Docs should render and components should behave as before with no regressions

Change-Id: I0e7633d8a6847993aa3f82e6b22a169f00a1f874
Reviewed-on: https://gerrit.instructure.com/c/instructure-ui/+/241041
Reviewed-by: Chris Guerrero <cguerrero@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Steve Jensen <sejensen@instructure.com>
Visual-Regression-Test: Steve Jensen <sejensen@instructure.com>
QA-Review: Ken Meleta <kmeleta@instructure.com>