chore: move generate:tokens command to bootstrap
We want to be able to run 'build' with the --scope argument, which we are unable to do if build is
executing the package build script and the tokens generation. The tokens generation is more similar
to the types generation and can be run on publish just like types.

TEST PLAN:
- Running 'yarn build --scope @instructure/ui-avatar' should execute without errors

Change-Id: Idbe132a66723fca796c02df7bcab0b323401dd91
Reviewed-on: https://gerrit.instructure.com/c/instructure-ui/+/241711
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ken Meleta <kmeleta@instructure.com>
QA-Review: Steve Jensen <sejensen@instructure.com>
Product-Review: Steve Jensen <sejensen@instructure.com>
Visual-Regression-Test: Steve Jensen <sejensen@instructure.com>