fix(esm): don't mark commonjs as esm
refs FOO-1331

when package.json specifies { "type": "module" }, files inside lib/
are treated by Node, starting from 12.5, as ESM when they are in fact
commonjs. This makes it impossible to require() them. To correctly
indicate the format to Node, this patch supplies the lib/ folder with a
package.json that specifies the intended format:

    { "type": "commonjs" }

:: test plan ::

using node 14, you are able to require a package that has both ESM and
CJS output:

    nvm use v14
    node ./packages/canvas-theme

Change-Id: Ibb58abf12a706b7fe63fa01723fc048b6913cfd7