refactor(emotion): make theme variables listed on docs page again
Closes: INSTUI-2823

Refactored componentTheme handling: the theme generator is now imported in `index.js` instead of
`styles.js`, and passed to and handled in `withStyle` decorator.
In the `withStyle` decorator we
expose the theme generator so we can use it in the docs client to generate the theme variables on
the component's page.
TEST PLAN:
The `Default Theme Variables` block appears on the docs page of migrated components.

refactor(ui-docs-client): remove unnecessary ApplyTheme import