refactor: remove hacks from test sandbox, fix broken test
Some tests were breaking because `displayName` was set in the themeable plugin in compile time. Luckily it turned out that simply using `name` is fine in the broken cases.

Also other tests were breaking with mysterious errors. Here the root cause was the `ui-test-sandbox` was doing some really weird stuff:

- Overriding Mocha's global exception handler (this was not working properly for some reason, maybe Mocha changed its internals?) to ignore some errors that were thrown.
- Overriding `console.error` to throw errors when a `console.error` message started with `Warning:`.

These were deleted.

As a result of deleting (1) one can see some harmless messages in the console when the tests are running
As a result of deleting (2) its no longer a valid strategy to surround `jsx` code with `try..catch` because writing something in the console does not cause errors to be thrown. Use `spy(console, 'error')` instead.

Closes: INSTUI-2877

BREAKING CHANGE:
It is no longer a valid strategy to surround `jsx` code with `try..catch` because writing something in the console does not cause errors to be thrown. Use `spy(console, 'error')` instead. Also, when accessing `Component.displayName`, use `Component.displayName || Component.name` with the name as a fallback instead.