fix(ui-checkbox): add spacing between single checkbox and messages
Closes: INSTUI-2808

Fixed bug: added spacing between the single checkbox and its error/info messages.
TEST PLAN:
When `messages` are passed to a single `Checkbox`, the messages are not supposed to be pressed up to
the label anymore.

BREAKING CHANGE:
VISUAL CHANGE: Since there is more space between the checkbox and the messages, it can potentially break layouts (vertically more pixels).