refactor(ui-link): refactor Link styling
Closes: INSTUI-2908

It was an error in the emotion migration to use `getElementType` to check if the link is a `button`
or an `a` tag, because sometimes we pass complex components as the `as` property. In this case the
elementType will give back a whole component, and the "elementType === 'button'" checks won't work
and apply the needed style. This is why we had to refactor the styling to use the css `&:is(button)`
selector, since it is applied as a css class style when the render is ready.