fix(ui-tree-browser): backport fix to remove line extending into root icon
Closes: INSTUI-2956

Backporting fix from PR #460: Tweaked the vertical line of "large" sized collections not to extend
into the collection icon.
Also tweaked the "small" sized collection to have the line ending closer
to the icon.
Added a size selector in the first example of TreeBrowser to be able to preview all
available sizes.