fix(ui-dialog): fix page scrolling up when Menu opens
Closes: INSTUI-3008

If a Dialog contains a focusable item, that item gets focused when the dialog opens. The browsers
automatically scroll the the focused element, but if the content and its css is not yet fully
loaded, the position of the focusable item might be not yet calculated, making the page jump to the
top. Now we wait a heartbeat before focusing the element. This fix affects other components using
`Dialog` under the hood: `DrawerLayout`, `Modal`, `Overlay`, `Popover` and `Tray`, and components
using these (e.g. Menu).