feat(ui-table): add `stackedSortByLabel` prop to `Table.ColHeader`
Closes: INSTUI-3116

The new `stackedSortByLabel` prop will allow users to set custom strings to the labels appearing in
the "sort by" select of a sortable table with stacked layout. It is using the `id` prop by default,
which is not always a user-friendly string. It will also allow to display longer, more descriptive
labels.

docs(ui-table): rephrase property description

Closes: INSTUI-3116