feat(ui-select,ui-simple-select): pass option props to the label before/after render prop
Closes: INSTUI-2903

If a function is passed to `renderBeforeLabel` or `renderAfterLabel` prop, the function will have a
`props` parameter. This way the the properties of that `Select.Option` or `SimpleSelect.Option`
(e.g.  `isDisabled`) is accessible in the render function, and can be used to set different colors
to that icon for different option state.