feat(ui-karma-config): tests can be run from the IDE
Now the little play buttons beside tests can be used to run and debug tests. The following changes
were needed:
-Removed the Webpack test loader plugin. This had an extra functionality that it was
adding a test to each package that just checked that it loads, this is gone.
-commented out the
non-working generateA11yTests() test, because it was silently throwing errors making debugging a
pain
-Replaced the DEBUG env var with NO_DEBUG that does the opposite, so tests are generated as
debug builds by default
-small changes to karma.config to get debugging to work
-slight refactor
to ReactComponentWrapper, so it doesnt spam warnings
TEST PLAN:
run tests in normal and watch mode, run a single test in these, try to debug a test in WebStorm and
in the browser