feat(ui-tabs): add 'fixHeight' prop to Tabs
Closes: INSTUI-3258

Refactored the DOM structure and CSS of Tabs component to have an extra wrapper for panels and use
flex layout under the hood. This makes it easier to handle the fix, min and max heights of the
component. Added a `fixHeight` prop to set a fix height, which can be set to `100%` now to fill the
parent element. Also added a few new examples.