chore: fix example generation
Fixed a bug where props where extracted wrong for components having sub-components attached on them
(e.g. Grid with GridCol and GridRow). The script used the last of these from an array, which was
always a sub-component, resulting in generating the prop combinations of the sub-component for the
main component. Changed it to use the first item of the array, which is the main component.