feat(ui-avatar): add `renderIcon` and `hasInverseColor` props to Avatar
Closes: INSTUI-3330

Instead of the initials, an icon can be displayed inside the Avatar with the new `renderIcon`
property. It worls with the `color` and `size` properties as well.
The `hasInverseColor` prop
inverts the background color and the text/icon color. Inverted Avatars have no border.
Add logic
that in case the image is unset in an update, show icons/initials again.
Backports PR #781