chore(ui-drilldown,shared-types,ui-options): new Drilldown component prototype
This commit contains the first version of the new Drilldown component. It is still WIP and does not contain documentation, tests, full a11y features and all examples.

Drilldown is supposed to be a diverse component that displays hierarchical data in a fashion that allows the users to "drill down" and dig deeper into more specifiy layers (pages) of the data structure. It is similar to the Menu, Select and TreeBrowser components.

Closes: INSTUI-3376