fix(ui-code-editor): change dynamically imported `CodeMirror` language modes
Closes: INSTUI-3394

our `CodeEditor` component dynamically imported `CodeMirror` language options
which was not
transpiled at all by `babel` which ment that we shipped
code that is not usable in browser
environments out of the box.
this change removes those dynamic `require`s and instead statically
imports them.
this might cause issues for users that uses this component and use jest in order to
test their components
I added notes on how to fix those jest related issues