refactor: fix test that were failing when running in random order
This commit fixes the following issues:
- We should not test for propTypes errors. PropTypes caches
these so they are displayed only once, so for example if 2 tests are checking for something like "id
is required in Button" then the second test will fail.
- there should be no code in "describe"
blocks (except constants), test code should go into it() functions
- make axe errors nicer, now it
displays the issue, its remedy and the violating code in one block.