feat(ui-drilldown,ui): add new `Drilldown` component
Closes: INSTUI-3446, INSTUI-3357

The `Drilldown` component exists to support navigating and managing tree structures in compact
spaces. It is a diverse component that displays hierarchical data in a fashion that allows the users
to “drill down” and dig deeper into the layers (pages) of the data structure. It has similar look
and features to the `Menu`, `Select` and `TreeBrowser` components.