feat(ui-range-input): add accessible variant for RangeInput handle
Closes: INSTUI-3462

Backported change from V8: the previous thumb design is not fully accessible (doesn't have
sufficiant color contrast and focus indication), so it has become deprecated. For this feature to be
non-breaking, we indtroduced a new `thumbVariant` prop on the component. The `deprecated` variant
(and connected theme variables) will be removed in V9, please use the `accessible` variant.