test(ui-billboard,ui-menu,ui-tabs): fix flaky VRT
Closes: INSTUI-3206

Chromatic tests are almost always failing on Tabs and Menu, because for some reason it cannot make the snapshots consistent, and the popovers and tab overflows keep repositioning. We1ve tried adding delays, but it didn't seem to have any effect. No we are trying additional paddings and repositioning.

Also added the missing `small` and `large` size version tests for Billboard.