feat(ui-tabs): expose `defaultOverflowY` theme variable on Tabs.Panel
Closes: INSTUI-3553

This theme variable allows the default `overflowY` to be set to `visible` on Tabs.Panel. It is
useful, when there is a 100% width interactive element (e.g. Button) in the panel, in which case the
focus ring would be cropped. Note, that this only works when the panel is not scrollable vertically.