feat(ui-truncate-list,ui): add `TruncateList` utility component
Closes: INSTUI-3599

TruncateList is a special kind of list component that can be configured to adapt to small screen
widths by truncating its items that don’t fit within the available space. It is a utility component
that can be used on its own, but its intended use is to handle the truncation logic of menu items in
navigation bars.