fix(ui-text-input): fix padding calculation for before/after element
Closes: INSTUI-3748

This is a workaround for an edge-case, when the TextInput's parent is hidden on load, so the element
is not visible either. In this case the computed width is going to be 'auto', so we assume it has
width so that the padding won't be removed. This fix affects all TextInput based components, e.g.:
Select, SimpleSelect, DateSelect, etc.