feat(ui-progress): add `shouldAnimate` prop to ProgressBar
Closes: INSTUI-3751

This change adds the `shouldAnimate` prop to ProgressBar, which makes animates the change between
values. To implement this feature, the complement had to be be refactored to \"hide\" the HTML
`\<progress\>` element in the DOM and render a custom progress bar in front of it. There are two
reasons for this: it was impossible to style the HTML progress to animate in Firefox as well, but we
had to keep it for accessibilty purposes.