fix(ui-selectable): fix Select options not being selectable on iOS Safari with VoiceOver on
aria-owns seems to be misused here. It rearranges the DOM seen by the screenreader, so the subtree
owned by the element will be its child. Backport of
https://github.com/instructure/instructure-ui/pull/1668
TEST PLAN:
Test all examples of Select, SimpleSelect, Selectable with keyboard navigation and VoiceOver. Test
on mobile devices with TalkBack/VoiceOver and Windows too if possible