fix(ui-select,ui-text-input): fix long before elements overflowing in TextInput, Select, SimpleSelect
Closes: INSTUI-4344

When adding lots of elements to renderBeforeInput for example Tags it was overflowing the input
field. Also when there are more than one line of elements keep the input field in the same line
if there is enough space
TEST PLAN:
Add lots of elements to renderBeforeInput to Select,SimpleSelect,TextInput. They should wrap,
not overflow