fix(ui-view): fix focus outline showing when setting a global CSS rule for outline
currently we are animating the focus ring on focus, on the normal state its a CSS outline with 0
alpha. If the outline is overridden it will be animated because its different from the global
default. This change fixes it by setting the :hover and :active outline states to the same
value as the default focus outline.

Fixes INSTUI-4567