refactor: remove jsdoc-api from the docs app
It uses the TypeScript compiler API instead of
JSDoc-api JSDoc is outdated and its meant for
 JavaScript projects.
 We use this to parse everything thats not a
 React component or a Markdown file.
 (the utilities menu in the sidebar).

It gets type information for functions from the
TypeScript types, so it can be deleted in the
jsdoc (the {Object} part in @param {Object} bla).

For non-function or complex utilities
it gets the info from the JSDoc so we need to
keep it (e.g. withStyle, contains).

Removed DateTime and conversions and some
custom PropTypes from the doc since they were
parsed in a special way and contained no useful
info.

note that it cannot parse types if the jsdoc is
at the export statement