fix(ui-modal,ui-dom-utils): fix Modal focus trap broken when it has a scrollbar
The issue was that its not recognized by findFocusable as a focusable element. Recently browsers
have made scrollable elements focusable, but there is no sure way to know that a scrollbar is
displayed, an easy workaround is to add the tabIndex=0 prop. Note that this breaks VoiceOver a bit,
its not able to navigate the inside of a scrollable container with SR focus

INSTUI-4625