fix(many): fix focus ring not respecting theme overrides in Button and FileDrop
This was already fixed in INSTUI-4375, but it broke with the focus rewrite. This PR introduces a
more stable method to accomplish the same thing. Instead of a new prop, now the borderRadius prop on
View accepts CSS units like '1rem'

To test:
- Set different borderRadius values in the View examples when it has a focus ring
- Override the borderRadius theme variable in the Button and FileDrop examples.
it should now respect the theme variable and not use the default value of the View.

Fixes INSTUI-4641