fix(ui-testable): handle ref forwarding in testable decorator
The testable decorator was passing ref as a regular prop instead of using
React's ref forwarding mechanism, causing React warnings when components
decorated with @testable() received a ref prop.

This fix updates the decorator to use React.forwardRef and filters out
the ref prop before passing props to the wrapped component, eliminating
the warning: "ref is not a prop. Trying to access it will result in
undefined being returned."