fix(ui-a11y-utils): fix focus region skipping items after Drilldown/scrollbars
The core of the issue is that findFocusable cannot detect everything that can be focused reliably,
e.g. browsers now focus on DOM elements that have scrollbars. The deleted code added the element
with the scrollbar to the end of the focus  region array, causing anything after it skipped when
tabbing. The downside of deleting this code is that focus order breaks when the last element in a
Popover/Modal is something that findFocusable does not detect e.g. a Menu (without a trigger,
 this is not detected because its tabIndex changes dynamically)

Fixes INSTUI-4694